Add more info about max_points to points API
Add fields max_points and max_points_by_difficulty to the points API
general information. This simplifies calculating relevant points in
MOOC-Jutut. The max_points field was already provieded in the API for
modules, but not for the course as a whole, and the max points for
each category/difficulty weren't in the API beforehand.